+++ src/app/DevilDaggersInfo.App/Ui/ImGuiExt.Hyperlink.cs
@@ -2,7 +2,7 @@ using ImGuiNET;
 using System.Diagnostics;
 using System.Numerics;
 
-namespace DevilDaggersInfo.App;
+namespace DevilDaggersInfo.App.Ui;
 
 public static partial class ImGuiExt
 {
